Finito il refactor, parte
