+++ src/main/java/backend/mapper/UserMapper.java
@@ -1,28 +1,21 @@
 package backend.mapper;
-
-import backend.dto.common.HasRole;
 import backend.dto.utente.CreateUserDTO;
 import backend.dto.utente.ResponseUserDTO;
 import backend.dto.utente.RoleAssignmentDTO;
 import backend.dto.utente.UpdateUserDTO;
 import backend.dto.utente.admin.CreateUserFromAdminDTO;
 import backend.dto.utente.admin.UpdateUserFromAdminDTO;
+import backend.mapper.common.UserRoleMappingSupport;
 import backend.mapper.resolver.RoleResolver;
 import backend.model.Ruolo;
 import backend.model.Utente;
 import backend.model.UtenteRuolo;
 import backend.repository.RuoloRepository;
 import org.mapstruct.*;
-import org.springframework.util.CollectionUtils;
-
-import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = MappingConstants.ComponentModel.SPRING, uses = {BenefitMapper.class, RoleMapper.class, RoleResolver.class, RuoloRepository.class})
-public interface UserMapper extends GenericMapper<Utente, CreateUserDTO, UpdateUserDTO, ResponseUserDTO> {
+public interface UserMapper extends GenericMapper<Utente, CreateUserDTO, UpdateUserDTO, ResponseUserDTO>, UserRoleMappingSupport {
 
     @Override
     @Mapping(
@@ -54,38 +47,6 @@ public interface UserMapper extends GenericMapper<Utente, CreateUserDTO, UpdateU
     @Mapping(target = "password", ignore = true)
     @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     Utente partialUpdateFromAdminUpdate(UpdateUserFromAdminDTO updateDto, @MappingTarget Utente utente);
-
-    @AfterMapping
-    default void linkRolesToUser(HasRole dto, @MappingTarget Utente utente, RuoloRepository ruoloRepository) {
-        // Usa il metodo dell'interfaccia
-        Set<UUID> ruoliIds = dto.ruoliIds();
-
-        if (CollectionUtils.isEmpty(ruoliIds)) {
-            if(utente.getUtenteRuoli() != null) {
-                utente.getUtenteRuoli().clear();
-            }
-            return;
-        }
-
-        List<Ruolo> ruoliTrovati = ruoloRepository.findAllById(ruoliIds);
-
-        Set<UtenteRuolo> newAssignments = ruoliTrovati.stream()
-                .map(ruolo -> {
-                    UtenteRuolo assignment = new UtenteRuolo();
-                    assignment.setUtente(utente);
-
-                    assignment.setRuolo(ruolo);
-                    return assignment;
-                })
-                .collect(Collectors.toSet());
-
-        if(utente.getUtenteRuoli() != null) {
-            utente.getUtenteRuoli().clear();
-            utente.getUtenteRuoli().addAll(newAssignments);
-        } else {
-            utente.setUtenteRuoli(newAssignments);
-        }
-    }
 
     default void updateRoleFromDto(RoleAssignmentDTO dto, Ruolo ruolo, @MappingTarget Utente utente) {
         if (dto == null || ruolo == null || utente == null) {
